# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API for Bigkinds AI web service.",
        "title": "Bigkinds AI",
        "contact": {}
    },
    "host": "gnew-biz.tplinkdns.com:8080",
    "paths": {
        "/v1/chats/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Create a new chat",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer upstage_kindsai_key",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "CreateChatRequest",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateChatRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/v1/chats/{chat_id}/completions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new Completion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Create a new Completion",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer upstage_kindsai_key",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "ffacea9b-d5a1-4844-8a0f-520b69a93ac3",
                        "description": "chat_id",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "CreateChatCompletionRequest",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateChatCompletionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/service.CreateChatCompletionResult"
                        }
                    }
                }
            }
        },
        "/v2/chats/{chat_id}/completions/multi": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new CompletionMulti",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Create a new CompletionMulti",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer upstage_kindsai_key",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "ffacea9b-d5a1-4844-8a0f-520b69a93ac3",
                        "description": "chat_id",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "CreateChatCompletionRequest",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateChatCompletionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/service.CreateChatCompletionResult"
                        }
                    }
                }
            }
        },
        "/v2/chats/{chat_id}/qas": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "채팅방 아이디를 전달하면 해당 채팅방 아이디에 해당하는 대화이력을 조회 함",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "채팅방의 대화 이력 조회",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer upstage_kindsai_key",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "ffacea9b-d5a1-4844-8a0f-520b69a93ac3",
                        "description": "chat_id",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "handler.CreateChatCompletionRequest": {
            "type": "object",
            "properties": {
                "job_group": {
                    "type": "string",
                    "example": "통계용"
                },
                "message": {
                    "type": "string",
                    "example": "트럼프 당선에 대해서 알려줘"
                },
                "provider": {
                    "type": "string",
                    "example": ""
                },
                "session": {
                    "type": "string",
                    "example": "session_id_value"
                }
            }
        },
        "handler.CreateChatRequest": {
            "type": "object",
            "properties": {
                "session": {
                    "type": "string"
                }
            }
        },
        "model.Completion": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "delta": {
                    "$ref": "#/definitions/model.CompletionDelta"
                },
                "id": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "token_usage": {
                    "type": "integer"
                }
            }
        },
        "model.CompletionDelta": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "references": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Reference"
                    }
                },
                "related_queries": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.Reference": {
            "type": "object",
            "properties": {
                "attributes": {
                    "$ref": "#/definitions/model.ReferenceAttributes"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "model.ReferenceAttributes": {
            "type": "object",
            "properties": {
                "byline": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "news_id": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "service.CreateChatCompletionResult": {
            "type": "object",
            "properties": {
                "completion": {
                    "$ref": "#/definitions/model.Completion"
                },
                "done": {
                    "type": "boolean"
                },
                "error": {}
            }
        }
    }
}